+++ Sources/SwiftSVG/Internal/Point+SwiftSVG.swift
@@ -2,7 +2,7 @@ import Swift2D
 
 extension Point {
     static var nan: Point {
-        return Point(x: Float.nan, y: Float.nan)
+        return Point(x: Double.nan, y: Double.nan)
     }
     
     var hasNaN: Bool {
@@ -10,12 +10,12 @@ extension Point {
     }
     
     /// Returns a copy of the instance with the **x** value replaced with the provided value.
-    func with(x value: Float) -> Point {
+    func with(x value: Double) -> Point {
         return Point(x: value, y: y)
     }
     
     /// Returns a copy of the instance with the **y** value replaced with the provided value.
-    func with(y value: Float) -> Point {
+    func with(y value: Double) -> Point {
         return Point(x: x, y: value)
     }
     
@@ -23,7 +23,7 @@ extension Point {
     ///
     /// This will explicitly check for `.isNaN`, and if encountered, will simply
     /// use the provided value.
-    func adjusting(x value: Float) -> Point {
+    func adjusting(x value: Double) -> Point {
         return (x.isNaN) ? with(x: value) : with(x: x + value)
     }
     
@@ -31,7 +31,7 @@ extension Point {
     ///
     /// This will explicitly check for `.isNaN`, and if encountered, will simply
     /// use the provided value.
-    func adjusting(y value: Float) -> Point {
+    func adjusting(y value: Double) -> Point {
         return (y.isNaN) ? with(y: value) : with(y: y + value)
     }
 }
