+++ Sources/SwiftSVG/Text.swift
@@ -10,24 +10,24 @@ import XMLCoder
 public struct Text: Element {
     
     public var value: String = ""
-    public var x: Float?
-    public var y: Float?
-    public var dx: Float?
-    public var dy: Float?
+    public var x: Double?
+    public var y: Double?
+    public var dx: Double?
+    public var dy: Double?
     
     // CoreAttributes
     public var id: String?
     
     // PresentationAttributes
     public var fillColor: String?
-    public var fillOpacity: Float?
+    public var fillOpacity: Double?
     public var fillRule: Fill.Rule?
     public var strokeColor: String?
-    public var strokeWidth: Float?
-    public var strokeOpacity: Float?
+    public var strokeWidth: Double?
+    public var strokeOpacity: Double?
     public var strokeLineCap: Stroke.LineCap?
     public var strokeLineJoin: Stroke.LineJoin?
-    public var strokeMiterLimit: Float?
+    public var strokeMiterLimit: Double?
     public var transform: String?
     
     // StylingAttributes
